chore(coin-evm): outsource helpers only consumed by Celo
