+++ libs/coin-modules/coin-celo/package.json
@@ -130,6 +130,7 @@
     "@ledgerhq/types-live": "workspace:^",
     "bignumber.js": "^9.1.2",
     "eip55": "^2.1.1",
+    "imurmurhash": "^0.1.4",
     "invariant": "^2.2.4",
     "lodash": "^4.17.21",
     "rxjs": "catalog:",
@@ -141,6 +142,7 @@
     "@ledgerhq/types-cryptoassets": "workspace:^",
     "@swc/core": "catalog:",
     "@swc/jest": "catalog:",
+    "@types/imurmurhash": "^0.1.4",
     "@types/invariant": "^2.2.37",
     "@types/jest": "catalog:",
     "@types/lodash": "^4.14.179",
