+++ libs/coin-modules/coin-evm/package.json
@@ -77,7 +77,6 @@
     "bignumber.js": "^9.1.2",
     "eip55": "^2.1.1",
     "ethers": "catalog:",
-    "imurmurhash": "^0.1.4",
     "invariant": "^2.2.2",
     "react-native-get-random-values": "^1.8.0"
   },
@@ -87,7 +86,6 @@
     "@ledgerhq/types-live": "workspace:^",
     "@swc/core": "catalog:",
     "@swc/jest": "catalog:",
-    "@types/imurmurhash": "^0.1.4",
     "@types/invariant": "^2.2.2",
     "@types/jest": "catalog:",
     "@types/node": "catalog:",
